requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=10.0.0
tqdm>=4.65.0
matplotlib>=3.7.0
scipy>=1.11.0
editdistance>=0.6.3
scikit-image>=0.21.0
einops>=0.7.0
sympy>=1.12
opencv-python>=4.8.0
tensorboard>=2.13.0
lxml>=4.9.0
regex>=2023.6.3
